pyproject.toml
[project]
name = "recipe-road"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "alembic>=1.16.1",
    "fastapi>=0.115.12",
    "pyaudio>=0.2.14",
    "pydantic-ai>=0.2.6",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "websocket-client>=1.8.0",
]
